requirements.txt
fastapi
uvicorn
psycopg[binary]
boto3
PyMuPDF
google-genai
Pillow
python-dotenv
colpali-engine
torch
pytest
